Organized components into folders and updated import paths
